+++ cmd/src/campaigns_common.go
@@ -201,8 +201,26 @@ func campaignsExecute(ctx context.Context, out *output.Output, svc *campaigns.Se
 	}
 	imageProgress.Complete()
 
+	pending = campaignsCreatePending(out, "Resolving repositories")
+	repos, err := svc.ResolveRepositories(ctx, campaignSpec)
+	if err != nil {
+		if repoSet, ok := err.(campaigns.UnsupportedRepoSet); ok {
+			campaignsCompletePending(pending, "Resolved repositories.")
+
+			block := out.Block(output.Line(" ", output.StyleWarning, "Some repositories are hosted on unsupported code hosts and will be skipped. Use the -allow-unsupported flag to avoid skipping them."))
+			for repo := range repoSet {
+				block.Write(repo.Name)
+			}
+			block.Close()
+		} else {
+			return "", "", errors.Wrap(err, "resolving repositories")
+		}
+	} else {
+		campaignsCompletePending(pending, "Resolved repositories.")
+	}
+
 	var progress output.Progress
-	specs, err := svc.ExecuteCampaignSpec(ctx, executor, campaignSpec, func(statuses []*campaigns.TaskStatus) {
+	specs, err := svc.ExecuteCampaignSpec(ctx, repos, executor, campaignSpec, func(statuses []*campaigns.TaskStatus) {
 		if progress == nil {
 			progress = out.Progress([]output.ProgressBar{{
 				Label: "Executing steps",