+++ internal/campaigns/errors.go
@@ -7,12 +7,12 @@ import (
 	"github.com/sourcegraph/src-cli/internal/campaigns/graphql"
 )
 
-// unsupportedRepoSet provides a set to manage repositories that are on
+// UnsupportedRepoSet provides a set to manage repositories that are on
 // unsupported code hosts. This type implements error to allow it to be
 // returned directly as an error value if needed.
-type unsupportedRepoSet map[*graphql.Repository]struct{}
+type UnsupportedRepoSet map[*graphql.Repository]struct{}
 
-func (e unsupportedRepoSet) Error() string {
+func (e UnsupportedRepoSet) Error() string {
 	repos := []string{}
 	typeSet := map[string]struct{}{}
 	for repo := range e {
@@ -32,10 +32,10 @@ func (e unsupportedRepoSet) Error() string {
 	)
 }
 
-func (e unsupportedRepoSet) appendRepo(repo *graphql.Repository) {
+func (e UnsupportedRepoSet) appendRepo(repo *graphql.Repository) {
 	e[repo] = struct{}{}
 }
 
-func (e unsupportedRepoSet) hasUnsupported() bool {
+func (e UnsupportedRepoSet) hasUnsupported() bool {
 	return len(e) > 0
 }